% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/assign.map.bins.R
\name{assign.map.bins}
\alias{assign.map.bins}
\title{Assign each place to a bin based on cutpoints}
\usage{
assign.map.bins(pctiles.vector, cutpoints = c((0:9)/10, 0.95, 1),
  labels = 1:11)
}
\arguments{
\item{pctiles.vector}{Vector of percentiles as decimal fractions, 0 to 1, one per place.}

\item{cutpoints}{Optional vector of cutpoints defining edges of bins. Default is every 0.10 from 0 to 1.00, as well as 0.95}

\item{labels}{vector of bin numbers, optional (default is 1 through 11, and NA values are put in bin 0).}
}
\value{
Returns a vector bin numbers.
}
\description{
For creating color-coded maps (choropleths), assign each place (row) to a bin.
Each bin represents one map color, and is defined by cutoff values.
}
\details{
The default bins 0-11 are defined as follows: \cr
bin 0: PCTILE=NA \cr
... \cr
bin 9:  0.80<=PCTILE<0.90 \cr
bin 10: 0.90<=PCTILE<0.95 \cr
bin 11: 0.95<=PCTILE<=1.00  \cr
}
\examples{
junk<-c(0,0.799,0.8, 0.8000001, 0.8999999,0.95,0.95001,1)
data.frame(pctile=junk, bin=assign.map.bins(junk))
# How it puts these in bins by default:
#      pctile bin   notes
#1  0.0000000   1
#2  0.7990000   8
#
#3  0.8000000   9 (0.80<=PCTILE<0.90 )
#4  0.8000001   9 (0.80<=PCTILE<0.90 )
#5  0.8999999   9 (0.80<=PCTILE<0.90 )
#
#6  0.9000000  10 (0.90<=PCTILE<0.95 )
#7  0.9001000  10 (0.90<=PCTILE<0.95 )
#8  0.9499990  10 (0.90<=PCTILE<0.95 )
#
#9  0.9500000  11 (0.95<=PCTILE<=1.00 ) I.E. THIS INCLUDES 100th percentile
#10 0.9500100  11 (0.95<=PCTILE<=1.00 ) I.E. THIS INCLUDES 100th percentile
#11 1.0000000  11 (0.95<=PCTILE<=1.00 ) I.E. THIS INCLUDES 100th percentile
}
\seealso{
\code{\link{make.bin.pctile.cols}} to call make.pctile.cols and make.bin.cols \cr
\code{\link{assign.pctiles}} to assign percentile (quantile) to each value within its column, in data.frame  \cr
\code{\link{assign.pctiles.alt2}} for alternative method to replicate assign.pctiles \cr
\code{\link{get.pctile}} to get (weighted) percentile of 1+ values within given vector of values \cr
\code{\link{assign.map.bins}} to assign map color bin using breaks to each value in data.frame given its percentile \cr
\code{\link{make.pctile.cols}} to take df and return same-sized df that is wtd.quantile of each value within its column \cr
\code{\link{make.pctile.cols.alt2}} to use alt version of make.pctile.cols for replication \cr
\code{\link{make.bin.cols}} to take df of pctile values and return same-sized df that is bin number (map color bin) using preset breaks. \cr
\code{\link{write.pctiles}} to save file that is lookup table of percentiles for columns of a data.frame \cr
\code{\link{write.pctiles.by.zone}} to save file that is lookup table of percentiles for columns of a data.frame, for each geographic zone (subset of rows) \cr
\code{\link{write.wtd.pctiles}} to save file that is lookup table of weighted percentiles for columns of a data.frame \cr
\code{\link{write.wtd.pctiles.by.zone}} to save file that is lookup table of weighted percentiles for columns of a data.frame, for each geographic zone (subset of rows) \cr
\code{\link{lookup.pctile}} to look up current approx weighted percentiles in a lookup table that is already in global memory \cr
}

