% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/assign.pctiles.R
\name{assign.pctiles}
\alias{assign.pctiles}
\title{Assign percentiles to values (weighted, by zone)}
\usage{
assign.pctiles(values, weights = NULL, zone = NULL, na.rm = TRUE)
}
\arguments{
\item{weights}{Optional, NULL by default (not fully tested), vector of weights for weighted percentiles (e.g., population weighted).}

\item{zone}{Optional, NULL by default, defines subsets of rows, so a percentile is found among rows that are within a given zone only.}

\item{na.rm}{NOT IMPLEMENTED HERE. Logical, optional, TRUE by default. Should NA values (missing data) be removed first to get percentile of those with valid data.
If FALSE, NA values are treated as being at the high percentiles.}

\item{x}{vector or data.frame}
}
\value{
Returns a numeric vector or data.frame same size as x, but if zone is specified, provides percentile with given zone.
}
\description{
For each column look at the distribution of values across all rows in a given zone (e.g., places in zone),
and find what percentile a given value is at.
}
\details{
Relies on the \code{\link{wtd.Ecdf}} function.
Could also add parameter like in rank(), na.last,
defining na.rm but also where to rank NA values if included, etc.
Default now is like na.last=NA, but like na.last='last' if na.rm=FALSE
Could also add parameter like in rank(), ties.method, defining if ties get min, max, or mean of percentiles initially assigned to ties.
Default for ties right now is like ties.method=max (which might not be what assign.pctiles() does in fact).
}
\examples{
x <- c(30, 40, 50, 12,12,5,5,13,13,13,13,13,8,9,9,9,9,9,10:20,20,20,20,21:30); wts <- rep(c(2,3), length(x)/2)
cbind(wts, x, PCTILE=assign.pctiles(x,wts))

# PERCENTILE OF ALL, NOT JUST THOSE WITH VALID DATA, IF na.rm=FALSE, but then NA values preclude high percentiles:
x <- c(NA, NA, NA, NA,NA,NA,NA,NA,NA,NA,13,13,8,9,9,9,9,9,10:20,20,20,20,21:30); wts <- rep(c(2,3), length(x)/2)
cbind(wts, x, PCTILE.alt2=assign.pctiles.alt2(x, wts, na.rm=FALSE), pctile=assign.pctiles(x,wts))[order(x),]
cbind(wts, x, PCTILE.alt2=assign.pctiles.alt2(x, wts, na.rm=TRUE), pctile=assign.pctiles(x,wts))[order(x),]

V=9
sum(wts[!is.na(x) & x <= V]) / sum(wts[!is.na(x)])

#A value (V) being at this PCTILE\% means that (assuming na.rm=TRUE):

# V >= x  for        PCTILE\% of wts     (for non-NA x), so
# V < x   for 100\% - PCTILE\% of wts     (for non-NA x), or
# PCTILE\% of all wts have V >= x (for non-NA x), so
# 100\% - PCTILE\% of all wts have V < x  (for non-NA x).

x <- c(32, NA, NA, NA,NA,NA,NA,NA,NA,NA,13,13,8,9,9,9,9,9,10:20,20,NA,20,21:30); wts <- rep(c(2,3), length(x)/2)
cbind(wts, x, PCTILE.alt2=assign.pctiles.alt2(x, wts, na.rm=FALSE), pctile=assign.pctiles(x,wts))[order(x),]
cbind(wts, x, PCTILE.alt2=assign.pctiles.alt2(x, wts, na.rm=TRUE), pctile=assign.pctiles(x,wts))[order(x),]
}
\seealso{
\code{\link{make.bin.pctile.cols}} to call make.pctile.cols and make.bin.cols \cr
\code{\link{assign.pctiles}} to assign percentile (quantile) to each value within its column, in data.frame  \cr
\code{\link{assign.pctiles.alt2}} for alternative method to replicate assign.pctiles \cr
\code{\link{get.pctile}} to get (weighted) percentile of 1+ values within given vector of values \cr
\code{\link{assign.map.bins}} to assign map color bin using breaks to each value in data.frame given its percentile \cr
\code{\link{make.pctile.cols}} to take df and return same-sized df that is wtd.quantile of each value within its column \cr
\code{\link{make.pctile.cols.alt2}} to use alt version of make.pctile.cols for replication \cr
\code{\link{make.bin.cols}} to take df of pctile values and return same-sized df that is bin number (map color bin) using preset breaks. \cr
\code{\link{write.pctiles}} to save file that is lookup table of percentiles for columns of a data.frame \cr
\code{\link{write.pctiles.by.zone}} to save file that is lookup table of percentiles for columns of a data.frame, for each geographic zone (subset of rows) \cr
\code{\link{write.wtd.pctiles}} to save file that is lookup table of weighted percentiles for columns of a data.frame \cr
\code{\link{write.wtd.pctiles.by.zone}} to save file that is lookup table of weighted percentiles for columns of a data.frame, for each geographic zone (subset of rows) \cr
\code{\link{lookup.pctile}} to look up current approx weighted percentiles in a lookup table that is already in global memory \cr
}

