% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgtest.R
\docType{data}
\name{bgtest}
\alias{bgtest}
\title{Test dataset for ejanalysis}
\usage{
data('bgtest', package = 'ejanalysis')
}
\description{
This data set provides test data, with Environmental, Demographic, and EJ Index fields,
 one row per block group, covering a few states. This is roughly a subset of ejscreen data (see \pkg{ejscreen} package).
}
\seealso{
\code{\link[ejscreen]{names.dvars}} \pkg{ejscreen}
}
\keyword{datasets}
