% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.fips1215.R
\name{clean.fips1215}
\alias{clean.fips1215}
\title{Check and clean Census block group or block FIPS - only check number of characters and if NA, not if valid number otherwise}
\usage{
clean.fips1215(fips)
}
\arguments{
\item{fips}{Vector of one or more elements, ideally character class, ideally 12 or 15 characters long (block group or block), required.}
}
\value{
Returns a vector of one or more character elements, same lengths as fips, NA if NA input
}
\description{
Check if valid Census block group or block FIPS, warn if not, add missing leading zero if inferred.
}
\details{
For information on FIPS codes, see \url{http://www.census.gov/geo/reference/ansi.html},
and also see \url{https://www.census.gov/geo/reference/geoidentifiers.html}
}
\examples{
  samplefips <- c("011030001003", "011030001003001", 11030001003001, 35, 1, NA, 'invalidtext', '02')
  clean.fips1215(samplefips)
  clean.fips1215("011030001003001")
}
\seealso{
\code{\link[=clean.fips]{clean.fips()}}
}
