% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ejanalysis.R
\name{ejanalysis}
\alias{ejanalysis}
\alias{ejanalysis-package}
\title{A package of tools for environmental justice (EJ) analysis}
\description{
This package provides tools for environmental justice (EJ) analysis,
including metrics characterizing disparities between demographic groups,
such as differences in environmental or other indicators measured for each Census unit
such as Census blocks, block groups, tracts, zip codes, or counties.
}
\concept{
justice EJ demographic
}

