% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geotype.R
\name{geotype}
\alias{geotype}
\title{Infer Type of Data as FIPS / ANSI Codes or Names of U.S. Geographies}
\usage{
geotype(x)
}
\arguments{
\item{x}{Required vector of one or more numeric or character FIPS and/or names of geographic locations.
Allowed types are State, County (or equivalent), tract, block group, and block.
Names for tracts, blockgroups, and blocks are not provided or interpreted.
FIPS codes here are all the relevant digits starting with the 2-character state FIPS,
so county fips must be 4-5 digits or characters for example (leading zeroes are inferred where
possible and included in outputs). See \code{\link{clean.fips}} for details.}
}
\value{
***TBD *** Returns *** types as character strings, and maybe cleaned values themselves? ***
}
\description{
Tries to interpret vector of one or more FIPS codes and/or names of geographies.
}
\examples{
# none yet
}
\seealso{
\code{\link{geofips}}, \code{\link[ejanalysis]{get.fips.st}} and related functions noted there, \code{\link[ejanalysis]{clean.fips}}, \code{\link[ejanalysis]{get.state.info}}
}

