% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.epa.region.R
\name{get.epa.region}
\alias{get.epa.region}
\title{Identify EPA Region for given state (or FIPS)}
\usage{
get.epa.region(state)
}
\arguments{
\item{state}{Vector of numeric or character class, required. Can be state FIPs as number or character, 2-character state abbreviation, or full state name}
}
\value{
Returns a vector of numbers, same length as state
}
\description{
Identify US Environmental Protection Agency region(s) containing given state(s) (or via FIPS)
}
\details{
For EPA Regions, see \url{http://www2.epa.gov/aboutepa#pane-4} or \url{http://www2.epa.gov/aboutepa/visiting-regional-office}.
For information on FIPS codes, see \url{http://www.census.gov/geo/reference/ansi.html},
and also see \url{https://www.census.gov/geo/reference/geoidentifiers.html}
}
\examples{
 myregions <- get.epa.region() # to see full list
 myregions <- get.epa.region('DC') # for one state by 2-letter abbreviation
 myregions <- get.epa.region(c('AK', 'NY', 'PR')) # for a vector of 2-letter abbreviations
 myregions <- get.epa.region(c('alaska', 'new york', 'puerto rico')) # vector of state names
 myregions <- get.epa.region(c('04', '36', '72')) # FIPS codes, character with leading zero
 myregions <- get.epa.region(c(4,36,72)) # state FIPS codes, numeric with no leading zeroes
 myregions <- get.epa.region(c('NY', 'Ohio')) # CANNOT MIX WAYS TO DEFINE STATES- PICK 1 FORMAT
 @seealso [clean.fips()]
}
