% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get.fips.bg.R
\name{get.fips.bg}
\alias{get.fips.bg}
\title{Get Census block group FIPS from longer FIPS}
\usage{
get.fips.bg(fips)
}
\arguments{
\item{fips}{Vector of one or more elements, character class, 12 or 15 characters long (block group or block), required.}
}
\value{
Returns a vector of one or more character elements, same lengths as fips
}
\description{
Extract partial FIPS code from longer FIPS code
}
\details{
Each fips passed to the function is a FIPS code (see \url{http://www.census.gov/geo/reference/ansi.html}.)
}
\examples{
get.fips.st("011030001003")
get.fips.st("011030001003001")
}

