% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.fips.county.R
\name{get.fips.county}
\alias{get.fips.county}
\title{Get County FIPS from block or block group FIPS}
\usage{
get.fips.county(fips, clean = TRUE)
}
\arguments{
\item{fips}{Vector of one or more elements, character class, 5, 11, 12 or 15 characters long after missing leading zeroes inferred (county, tract, block group or block), required.}

\item{clean}{Does not use clean.fips() if FALSE, which helps if the countiesall or other list is not yet updated, for example and lacks some new FIPS code}
}
\value{
Returns a vector of one or more character elements, same lengths as fips
NA where input cannot be converted to county/tract/block group/block length code, as with state fips as input.
}
\description{
Extract partial FIPS code from longer FIPS code
}
\details{
Each fips passed to the function is a FIPS code (see \url{http://www.census.gov/geo/reference/ansi.html}.)
It does not check to see if the codes are valid other than counting how many characters each has if clean=FALSE.
}
\examples{
samplefips <- c("011030001003", "011030001003001", 02610, 11030001003001, 35, 1,
 NA, 'invalidtext', '02')
get.fips.county(samplefips)
}
\seealso{
\code{\link[=clean.fips]{clean.fips()}}
}
