% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.fips.etc.R
\name{get.fips.etc}
\alias{get.fips.etc}
\title{Get tract, county, state, region info from US Census block group FIPS codes}
\usage{
get.fips.etc(fips)
}
\arguments{
\item{fips}{Vector of US Census block group FIPS codes (missing leading zeroes are added).}
}
\value{
Returns a data.frame with these fields: c('FIPS', 'FIPS.TRACT', 'FIPS.COUNTY', 'FIPS.ST', 'ST', 'countyname', 'statename', 'REGION')
  where FIPS is the input fips, the next few are the first few characters of fips corresponding to tract, county, or state, ST is the 2-letter state abbreviation, statename is state name, countyname is county name, and REGION is USEPA Region 1-10.
}
\description{
Use US Census block group FIPS codes to get more information about each block group's overall location, such as State and County name.
  It does not check to see if the codes are valid other than counting how many characters each has.
}
\examples{
 x=c("391670211002", "060730185143", "261079609003", 02, 02610,
  "400353734002", "371190030121", "250235022001", "550439609001", "060730170302")
 get.fips.etc(x)
}
