% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.fips.ST.R
\name{get.fips.st}
\alias{get.fips.st}
\title{Get State FIPS from block or block group FIPS}
\usage{
get.fips.st(fips, clean = TRUE)
}
\arguments{
\item{fips}{Vector of one or more elements, character class, 12 or 15 characters long (block group or block), required.}

\item{clean}{Does not use clean.fips() if FALSE, which helps if the countiesall or other list is not yet updated, for example and lacks some new FIPS code}
}
\value{
Returns a vector of one or more character elements, same lengths as fips
}
\description{
Extract partial FIPS code from longer FIPS code
}
\details{
For information on FIPS codes, see \url{http://www.census.gov/geo/reference/ansi.html},
and also see \url{https://www.census.gov/geo/reference/geoidentifiers.html}
It does not check to see if the codes are valid other than counting how many characters each has.
}
\examples{
samplefips <- c("011030001003", "011030001003001", 02610, 11030001003001, 35, 1,
 NA, 'invalidtext', '02')
get.fips.st(samplefips)
}
\seealso{
\code{\link[=clean.fips]{clean.fips()}}
}
