% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junk.R
\name{junk}
\alias{junk}
\title{Test dataset for ejanalysis}
\usage{
junk(a, b, c = 17)
}
\description{
This data set provides test data, with Environmental, Demographic, and EJ Index fields,
 one row per block group, covering a few states. This is roughly a subset of ejscreen data (see \pkg{ejscreen} package).
 Can be used like this:  data('bgtest', package = 'ejanalysis')
}
\seealso{
\code{\link[ejscreen]{names.dvars}} \pkg{ejscreen}
}
