% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make.bin.pctile.cols.byzone.R
\name{make.bin.pctile.cols.byzone}
\alias{make.bin.pctile.cols.byzone}
\title{Percentile and bin number fields, by zone, such as percentiles within each State}
\usage{
make.bin.pctile.cols.byzone(bg, zone, wtsvarname, keyvarname = "FIPS",
  datavarnames = gsub(keyvarname, "", names(bg)), ...)
}
\arguments{
\item{bg}{Data.frame of data field(s), weights (optional), and a unique ID (e.g., FIPS), one row per place if geographic data is used.}

\item{zone}{Optional vector specifying subsets of rows (e.g. State name) within which percentiles are calculated}

\item{wtsvarname}{Name of field in bg that has optional weights for weighted percentiles. Default is unweighted.}

\item{keyvarname}{Name of field in bg that has unique ID per row (per place). Default is "FIPS"}

\item{datavarnames}{Vector of names of fields in bg that have numeric data for which percentiles are calculated.}

\item{...}{Other parameters passed to \code{\link{make.bin.pctile.cols}} such as \code{cutpoints} and \code{labels}}
}
\value{
Returns a data.frame with one percentiles field and one bin numbers field for each of the datavarnames
}
\description{
Get weighted percentiles and bin numbers, by subset of rows (zone), for each column of data.
  This can be used to calculate population-weighted percentiles within each State, for national data on Census block groups, for example.
}

