% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.pctile.cols.R
\name{make.pctile.cols}
\alias{make.pctile.cols}
\title{Make columns of (weighted) percentiles from columns of values}
\usage{
make.pctile.cols(
  values,
  weights = 1,
  prefix = "pctile.",
  as.df = TRUE,
  zone = NULL
)
}
\arguments{
\item{values}{Data.frame, matrix, or vector of values. One column at a time is analyzed using \code{\link[=assign.pctiles]{assign.pctiles()}}}

\item{weights}{Optional numeric vector of weights (default is unweighted) to create weighted percentiles, such as population-weighted quantiles. Unweighted if not specified. Vector same length as number of rows in values parameter (or elements if vector).}

\item{prefix}{Optional character element, default is 'pctile.', provides text to paste to beginning of input data.frame column names to use as output column names.}

\item{as.df}{Optional logical TRUE by default, in which case matrix results are converted to data.frame}

\item{zone}{NULL by default, but if a vector is provided, it defines zones to group by, so percentiles are within a given zone only.}
}
\value{
Returns a matrix or data.frame
}
\description{
Takes values (vector, matrix, or data.frame) and returns percentiles,
showing the percentile of a value within all values in its column.
}
\seealso{
\code{\link[=make.bin.pctile.cols]{make.bin.pctile.cols()}} and \code{\link[=assign.pctiles]{assign.pctiles()}}
}
