% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/names.dvars.R
\docType{data}
\name{names.dvars}
\alias{Dlist}
\alias{demographic-variables}
\alias{names.d}
\alias{names.d.subgroups}
\alias{names.dvars}
\title{Fieldnames of demographic columns in ejanalysis data}
\format{A series of variables (each is a character vector of colnames):
\itemize{
 \item "names.d" (VSI.eo, VSI.svi6, pctmin, pctlowinc, pctlths, pctlingiso, pctunder5, pctover64)
 \item "names.d.bin"
 \item "names.d.eo"
 \item "names.d.eo.bin"
 \item "names.d.eo.pctile"
 \item "names.d.pctile"
 \item "names.d.subgroups"
 \item "names.d.subgroups.count"
 \item "names.d.subgroups.pct"
 \item "names.d.svi6"
 \item "names.d.svi6.bin"
 \item "names.d.svi6.pctile"
 #'  \item "Dlist" (this one is like names.d, but as a list, not a vector)
}}
\source{
Names developed for this package. No external data source.
}
\usage{
data('names.dvars')
}
\description{
This data set provides variables that hold the colnames
  of demographic fields in data.frames that may be used in the ejanalysis package
  to make it easier to refer to them as a vector, e.g., mydf[ , names.e]
}
\keyword{datasets}

