% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.ej.R
\docType{data}
\name{names.ej}
\alias{names.ej}
\alias{EJ-variable-names}
\title{Fieldnames of EJ index columns in ejscreen data}
\description{
This data set provides variables that hold the colnames
of environmental justice index fields in data.frames that may be used in the ejscreen package
to make it easier to refer to them as a vector
}
\seealso{
\link[ejscreen:names.d]{ejscreen::names.d}  \link[ejscreen:names.e]{ejscreen::names.e}  \link[ejscreen:names.ej]{ejscreen::names.ej}
}
\keyword{datasets}
