% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.ejvars.R
\docType{data}
\name{names.ejvars}
\alias{names.ejvars}
\alias{names.ej}
\alias{EJ-variable-names}
\title{Fieldnames of EJ index columns in ejscreen data}
\description{
This data set provides variables that hold the colnames
  of environmental justice index fields in data.frames that may be used in the ejscreen package
  to make it easier to refer to them as a vector, e.g., mydf[ , names.ej] \cr
  data('names.ejvars', package = 'ejscreen')
}
\seealso{
\code{\link[ejscreen]{names.ejvars}}
}
\keyword{datasets}
