% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pct.moe.R
\name{pct.moe}
\alias{pct.moe}
\title{Margin of Error for a Percent (Ratio)}
\usage{
pct.moe(a, b, a.moe, b.moe)
}
\arguments{
\item{a}{Numerators, as a vector}

\item{b}{Denominators, as a vector (should be same length as a, or it is recycled)}

\item{a.moe}{Margins of error for numerators, as a vector  (should be same length as a).}

\item{b.moe}{Margins of error for denominators, as a vector (should be same length as a, or it is recycled).}
}
\value{
Returns margin(s) of error same shape as parameter a
}
\description{
Estimates the margin of error (MOE) that characterizes uncertainty, for a ratio (a/b),
  based on a, b, and their MOE values.
}
\details{
This is based on US Census Bureau recommendations for working with American Community Survey summary file data.
  See \url{http://www.census.gov/programs-surveys/acs/guidance.html} and \url{http://www.census.gov/library/publications/2009/acs/researchers.html}
  For example, one can estimate MOE for percent poor in a block group, given
  estimates and margins of error for the count who are poor, and the count for whom poverty status is known.
}
\examples{
	 x <- pct.moe(15, 100, 3, 10)
}
\seealso{
\code{\link{sumoe}}
}
