% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop.ecdf.dd.R
\name{pop.ecdf.dd}
\alias{pop.ecdf.dd}
\title{Draw an Ecdf plot showing the distribution of scores in each demographic subgroup}
\usage{
pop.ecdf.dd(
  e,
  elab,
  mydf,
  dlabs = colnames(mydf),
  mywt = rep(1, NROW(mydf)),
  main,
  mycolors,
  refcolor = "black",
  sorted = TRUE,
  log = "",
  ...
)
}
\arguments{
\item{e}{Required numeric vector of scores (e.g., environmental indicator values) for places like block groups, one place per row}

\item{elab}{Required character single element vector with text label for environmental score, used in xlab of plot}

\item{mydf}{Required data.frame where each row is a place and each column is the percent of place population that is in given demographic subgroup, one subgroup per column}

\item{dlabs}{Optional, charater vector that specifies text for legend, same length as number of columns in mydf}

\item{mywt}{Optional, default is unweighted, but normally would specify as the population counts for places analyzed}

\item{main}{Optional character specifying plot title at top of plot. Default title notes names of subgroups}

\item{mycolors}{Optional vector of colors as long as number of columns in mydf, specifies colors of lines}

\item{refcolor}{Optional color, default is 'black'}

\item{sorted}{Optional logical, default is TRUE. Should the legend be sorted in order of population weighted mean values of e?}

\item{log}{Default is ''. Passed to \code{\link[=pop.ecdf]{pop.ecdf()}}}

\item{...}{Other optional parameters to pass to Ecdf}
}
\value{
Draws a plot
}
\description{
Draws a plot using \code{\link[Hmisc:Ecdf]{Hmisc::Ecdf()}}, overlaying cumulative distribution functions,
one for each subgroup specified, and one for overall population.
Useful to compare subgroups based on the distribution of scores among people in that subgroup
using data from small places like census block groups, based on having for each place the pop total
and \% of pop that is in each group.
}
\examples{
## #
\dontrun{
set.seed(99)
pctminsim=c(runif(7000,0,1), pmin(rlnorm(5000, meanlog=log(0.30), sdlog=1.7), 4)/4)
popsim= runif(12000, 500, 3000)
esim= rlnorm(12000, log(10), log(1.15)) + rnorm(12000, 1, 0.5) * pctminsim - 1
}
}
\seealso{
\code{\link[Hmisc:Ecdf]{Hmisc::Ecdf()}} \code{\link[=RR]{RR()}} \code{\link[=pop.cdf]{pop.cdf()}}   \code{\link[=pop.cdf2]{pop.cdf2()}} \code{\link[=pop.ecdf]{pop.ecdf()}}  \code{\link[=pop.cdf.density]{pop.cdf.density()}}
}
