% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pop.ecdf.R
\name{pop.ecdf}
\alias{pop.ecdf}
\title{Draw an Ecdf plot comparing distributions of scores in selected demographic groups}
\usage{
pop.ecdf(scores, pcts, pops, allothers = TRUE, col = "red", main, weights,
  subtitles = FALSE, ...)
}
\arguments{
\item{scores}{Numeric vector, required. Values to analyze.}

\item{pcts}{Numeric vector or data.frame, required. Same number of vector elements or data.frame
rows as length of scores. Specifies the fraction of population that is in demographic group(s) of interest, one row per place.}

\item{pops}{***}

\item{allothers}{Logical value, optional, TRUE by default. ***}

\item{col}{Optional, default is 'red' to signify line color red for key demographic group}

\item{main}{Required character specifying plot title}

\item{weights}{Vector of weights for weighted frequency distributions}

\item{subtitles}{Logical FALSE by default}

\item{...}{other optional parameters to pass to Ecdf}
}
\value{
Same as \code{\link[Hmisc]{Ecdf}} -- draws a plot
}
\description{
Draws a plot using \code{\link[Hmisc]{Ecdf}}, overlaying cumulative distribution functions,
 one for each subgroup specified.
 Useful to compare 2 groups based on each groups entire pdf or cdf  distribution of peoples scores,
 using data from small places like census block groups, based on having for each place the pop total
 and % of pop that is in each group or perhaps already have count in each group.
}
\details{
Notes: \cr
to compare zones, \cr
compare demog groups, (see parameter called group)\cr
compare multiple groups and/or multiple zones, like hisp vs others in us vs ca all on one graph\cr
see \code{\link{Ecdf}} for options & try passing a data.frame instead of just vector\cr
#' \cr\cr
}
\examples{
###
\dontrun{
# pop.ecdf( 31:35, c(0.10, 0.10, 0.40, 0, 0.20), 1001:1005 )
# pop.ecdf(dat$Murder, dat$Population * (dat$Illiteracy/100))
#
# pop.ecdf(places$pm, places$pctmin, places$pop)
# pop.ecdf(log10(places$traffic.score), places$pctmin, places$pop)
# pop.ecdf(places$cancer, places$pctmin, places$pop, allothers=FALSE); pop.ecdf(places$cancer, places$pctlingiso, places$pop, col='green', allothers=FALSE, add=TRUE)
# Demog suscept  for each REGION (can't see if use vs others)
pop.ecdf(bg$traffic.score, bg$VSI.eo, bg$pop, log='x', subtitles=FALSE,
         group=bg$REGION, allothers=FALSE,
         xlab='Traffic score (log scale)', ylab='\%ile of population', main='Distribution of scores by EPA Region')

# Demog suscept (how to show vs others??), one panel per REGION
pop.ecdf(bg[ , names.e], bg$VSI.eo, bg$pop, log='x', subtitles=FALSE,
         allothers=TRUE, ylab='\%ile of population', main='Distribution of scores by EPA Region')

# log scale is useful & so are these labels passed to function
# in CA vs not CA
pop.ecdf(bg$traffic.score, bg$ST=='CA', bg$pop,
         subtitles=FALSE,
         log='x', xlab='\%ile of population', ylab='Traffic scores (log scale)',
         main='Distribution of scores in CA (red) vs rest of US')

# Flagged vs not (all D, all zones)
pop.ecdf(bg$traffic.score, bg$flagged, bg$pop, log='x')

# D=Hispanics vs others, within CA zone only
pop.ecdf(bg$traffic.score, bg$ST=='CA', bg$pop * bg$pcthisp, log='x')
# Demog suscept vs others, within CA only
pop.ecdf(bg$traffic.score, bg$ST=='CA', bg$pop * bg$VSI.eo, log='x')

}
}
\seealso{
\code{\link[Hmisc]{Ecdf}} \code{\link{RR}}
}

