% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollup.pct.R
\name{rollup.pct}
\alias{rollup.pct}
\title{Calculate a/b for each Subset}
\usage{
rollup.pct(a, b, zone)
}
\arguments{
\item{a}{Required numeric vector, numerator}

\item{b}{Required numeric vector, denominator. Same length as a.}

\item{zone}{Optional, vector to group by. Same length as a and b.}
}
\value{
Returns a table with a/b calculated within each zone.
}
\description{
Uses data.table package to quickly calculate
  ratio of a/b within each subset of a dataset (e.g., by zone).
  This will be superseded by \code{\link{rollup}} once that is completed.
}
\examples{
  pre1960=1:100; builtunits=rep(c(10, 100),50); zone=rep(c('NY','MA'),50)
  rollup.pct(a,b,zone)
}
\seealso{
\code{\link{make.bin.pctile.cols}} to call functions below, converting columns of values to percentiles and then bins \cr
\code{\link{assign.pctiles}} for one vector, assign (weighted) percentile (quantile) to each value within its zone (subset)  \cr
\code{\link{assign.pctiles.alt2}} as an alternative method, to replicate assign.pctiles, but not by zone \cr
\code{\link{get.pctile}} to get (weighted) percentile of just 1+ values within given vector of values \cr
\code{\link{make.pctile.cols}} for a data.frame, assign percentiles, return a same-sized df that is wtd.quantile of each value within its column \cr
\code{\link{make.pctile.cols.alt2}} as an alternative method, to replicate make.pctile.cols \cr
\code{\link{assign.map.bins}} for one vector (or data.frame) of values (e.g. percentiles), return same-sized df that is bin number (map color bin) using preset breaks. \cr
\code{\link{make.bin.cols}} for a data.frame of values (e.g. percentiles), return same-sized df that is bin number (map color bin) using preset breaks. \cr
\code{\link{write.pctiles}} to save file that is lookup table of percentiles for columns of a data.frame \cr
\code{\link{write.pctiles.by.zone}} to save file that is lookup table of percentiles for columns of a data.frame, for each geographic zone (subset of rows) \cr
\code{\link{write.wtd.pctiles}} to save file that is lookup table of weighted percentiles for columns of a data.frame \cr
\code{\link{write.wtd.pctiles.by.zone}} to save file that is lookup table of weighted percentiles for columns of a data.frame, for each geographic zone (subset of rows) \cr
\code{\link{lookup.pctile}} to look up current approx weighted percentiles in a lookup table that is already in global memory \cr
}
