% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RR.plotbar.R
\name{RR.plotbar}
\alias{RR.plotbar}
\title{Draw barplot comparing mean score in a group vs reference group}
\usage{
RR.plotbar(
  x,
  dname,
  ename,
  dlab = dname,
  elab = ename,
  reflab = paste("Non-", dlab, sep = ""),
  cex.names = 1.9,
  cex.axis = 1.5,
  ...
)
}
\arguments{
\item{x}{Results of \code{\link[=RR.means]{RR.means()}} function, mean indicator score or risk by demographic group}

\item{dname}{Demographic group name to be found among names of first dim of x. Required.}

\item{ename}{Environmental factor or risk type name to be found as names of second dim of x. Required.}

\item{dlab}{optional character vector of label to use instead of dname on the plot. Default is dname.}

\item{elab}{optional character vector of label to use instead of ename on the plot. Default is ename.}

\item{reflab}{optional character vector of lable to use for reference group. Default is Non-dlab (e.g., "Non-Poor")}

\item{cex.names}{optional numeric vector passed to \code{\link[=barplot]{barplot()}}}

\item{cex.axis}{optional numeric vector passed to \code{\link[=barplot]{barplot()}}}

\item{...}{optional other parameters passed to \code{\link[=barplot]{barplot()}}}
}
\value{
draws a barplot, returns the two values as named list
}
\description{
Draws a plot using group means from \code{\link[=RR.means]{RR.means()}}
}
\examples{
\dontrun{
    data(bgtest, package = 'ejanalysis')
    ### x should be the output of RR.means()
    x <- RR.means(e = bgtest$traffic.score, d = bgtest$pctmin, pop = bgtest$pop)
    RR.plotbar(x, 'pctlths', 'proximity.rmp')
    RR.plotbar(x, 'pctmin', 'proximity.tsdf', dlab = 'Minorities',
      elab = 'Haz. Waste TSD Facility Proximity Score', ylab = 'Proximity Score')
 }
}
\seealso{
\code{\link[=RR.plot]{RR.plot()}}
}
