% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RR.table.add.R
\name{RR.table.add}
\alias{RR.table.add}
\title{Merge table of Relative Risk results for some zones with table for USA overall}
\usage{
RR.table.add(rrs1, rrs2, zones2)
}
\arguments{
\item{xxx}{xxx/what/type/default.}

\item{xxx}{xxx/what/type/default.}

\item{xxx}{xxx/what/type/default.}

\item{xxx}{xxx/what/type/default.}
}
\value{
Returns a xxx/type/what/size.
}
\description{
xxx.
}
\details{
xxx.
Returns a xxx.
xxx.
}
\note{
Future work: xxx.
}
\examples{
RRS.US  <- RR.table(mydat=bg, Enames=names.e, Dnames=c(names.d, names.d.subgroups.pct), popcolname='pop')
RRS.ST  <- RR.table(mydat=bg, Enames=names.e, Dnames=c(names.d, names.d.subgroups.pct), popcolname='pop', Zcolname='ST')
RRS <- RR.table.add(RRS.ST, RRS.US)
RRS[ 'pctlowinc', , ]
RRS[ , , 'CA']
RRS[ , 'pm', ]
RRS.REGION  <- RR.table(mydat=bg, Enames=names.e, Dnames=c(names.d, names.d.subgroups.pct), popcolname='pop', Zcolname='REGION')
RRS2 <- RR.table.add(RRS, RRS.REGION)
RRS2[ , , '8']
}
\seealso{
xxx to do xxx.
}
\keyword{EJ}

