% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RR.table.addmaxzone.R
\name{RR.table.addmaxzone}
\alias{RR.table.addmaxzone}
\title{Add to an existing RR.table the max relative risk value from any zone in the table}
\usage{
RR.table.addmaxzone(x, testing = FALSE)
}
\arguments{
\item{x}{A table of relative risk values from RR.table()}

\item{testing}{If TRUE, print extra info to terminal}
}
\description{
Add to an existing RR.table the max relative risk value from any zone in the table
}
