% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RR.table.max.R
\name{RR.table.max}
\alias{RR.table.max}
\title{See which group and risk indicator (in zone) have max values in RR table}
\usage{
RR.table.max(x, by = 1, digits = 3)
}
\arguments{
\item{x}{Array results from \code{\link[=RR.table]{RR.table()}} or related function}

\item{by}{optional vector (numeric or character), specifies which of three dimensions should be shown.
Default is 1, meaning dimension 1 which is demographic indicator.}

\item{digits}{optional number of digits to round to, default is 3.}
}
\value{
The full x array, except only the first two elements of the specified dimension,
which should be the two elements with the max number and text saying which name is that max.
}
\description{
See which Demographic group has highest RR for each Envt risk indicator,
and which Envt risk indicator has highest RR for each Demographic group,
within each zone (USA, etc.),
in table of Relative Risk results by zone by group by envt risk factor.
}
\seealso{
\itemize{
\item \code{\link{ej.indexes}} for local contribution to a variety of overall disparity metrics such as excess risk
\item \code{\link{RR}} to calculate overall disparity metric as relative risk (RR), ratio of mean environmental indicator values across demographic groups
\item \code{\link{RR.table}} to create 3-D table of RR values, by demographic group by environmental indicator by zone
\item \code{\link{RR.table.sort}} to sort existing RR table
\item \code{\link{RR.table.add}} to add zone(s) to existing RR table
\item \code{\link{write.RR.tables}} to write a file with a table or RR by indicator by group
\item \code{\link{ej.added}} to find EJ Index as local contribution to sum of EJ Indexes
\item \code{\link{RR.cut.if.gone}} to find local contribution to RR
\item \code{\link{RR.if.address.top.x}} to find how much RR would change if top-ranked places had different conditions
\item \code{\link{rrfv}} for obsolete attempt to vectorize rrf
\item \code{\link{rrf}} for older simpler function for RR of one indicator for one demographic group
\item \code{\link{pop.ecdf}} to compare plots of cumulative frequency distribution of indicator values by group
}
}
