% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RR.table.sort.R
\name{RR.table.sort}
\alias{RR.table.sort}
\title{Sort table of Relative Risk results by zone by group by envt risk factor}
\usage{
RR.table.sort(x, margins = c(1, 2, 3), decreasing = TRUE)
}
\arguments{
\item{x}{Array of RR values from \code{\link{RR}} in array of 3 dimensions: x[Dnames, Enames, Zcolnames]}

\item{margins}{optional numeric vector of which dimensions to sort, default is all 3, so margins = c(1, 2, 3) by default}

\item{decreasing}{default is TRUE, defines how to sort (note all specified dimensions are sorted the same way by this function)}
}
\value{
Sorted version of x array of 3 dimensions: RRS[Dnames, Enames, Zcolnames]
  Returns an array with
  one demographic group per row,
  one environmental risk indicator per column, and
  third dimension for which zone (e.g., which US State)
}
\description{
Sort table of Relative Risk results by zone by group by envt risk factor
}
\examples{
RRS.US  <- RR.table(mydat=bg, Enames=names.e, Dnames=c(names.d, names.d.subgroups.pct),
 popcolname='pop')
RRS.ST  <- RR.table(mydat=bg, Enames=names.e, Dnames=c(names.d, names.d.subgroups.pct),
 popcolname='pop', Zcolname='ST')
RRS <- RR.table.add(RRS.ST, RRS.US)
RRS[ 'pctlowinc', , ]
RRS[ , , 'CA']
RRS[ , 'pm', ]

RR.plot(RR.table.sort(RRS))
RR.table.sort(RRS, 1)

RRS.REGION  <- RR.table(mydat=bg, Enames=names.e, Dnames=c(names.d, names.d.subgroups.pct),
 popcolname='pop', Zcolname='REGION')
RRS2 <- RR.table.add(RRS, RRS.REGION)
RRS2[ , , '8']
}
\seealso{
\itemize{
\item \code{\link{ej.indexes}} for local contribution to a variety of overall disparity metrics such as excess risk
\item \code{\link{RR}} to calculate overall disparity metric as relative risk (RR), ratio of mean environmental indicator values across demographic groups
\item \code{\link{RR.table}} to create 3-D table of RR values, by demographic group by environmental indicator by zone
\item \code{\link{RR.table.sort}} to sort existing RR table
\item \code{\link{RR.table.add}} to add zone(s) to existing RR table
\item \code{\link{write.RR.tables}} to write a file with a table or RR by indicator by group
\item \code{\link{ej.added}} to find EJ Index as local contribution to sum of EJ Indexes
\item \code{\link{RR.cut.if.gone}} to find local contribution to RR
\item \code{\link{RR.if.address.top.x}} to find how much RR would change if top-ranked places had different conditions
\item \code{\link{rrfv}} for obsolete attempt to vectorize rrf
\item \code{\link{rrf}} for older simpler function for RR of one indicator for one demographic group
\item \code{\link{pop.ecdf}} to compare plots of cumulative frequency distribution of indicator values by group
}
}
\keyword{EJ}
