% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RR.table.vs.us.R
\name{RR.table.vs.us}
\alias{RR.table.vs.us}
\title{Relative Risks (RR) in zones as ratios to US values}
\usage{
RR.table.vs.us(
  x,
  d = names.d[names.d \%in\% names(x)],
  e = names.e[names.e \%in\% names(x)],
  zone = names(x[1, 1, ])
)
}
\arguments{
\item{x}{A three dimensional array that is created by \code{\link[=RR]{RR()}} and has format like RRS\link{d, e, zone} but where zone must include the name USA}

\item{d}{Vector of names of demographic groups that must be subset of first dimension of RRS}

\item{e}{Vector of names of environmental risk indicators that must be subset of second dimension of RRS}

\item{zone}{Vector of zone names that must be subset of names of third dimension of RRS}
}
\value{
numeric results same shape as RRS\link{d, e, zone}
}
\description{
Takes zone-specific RRS values and divides each value by the corresponding value for the USA overall.
}
\seealso{
\code{\link[=RR]{RR()}}
}
