% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotratios.R
\name{sim_plotratios}
\alias{sim_plotratios}
\title{create plot of ratios indicating disparity, from sim_riskbydistance output}
\usage{
sim_plotratios(x)
}
\arguments{
\item{x}{output from sim_riskbydistance}
}
\description{
create plot of ratios indicating disparity, from sim_riskbydistance output
}
\seealso{
sim_riskbydistance
}
