% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotrisk.R
\name{sim_plotrisk}
\alias{sim_plotrisk}
\title{create plot of basic info from sim_riskbydistance output}
\usage{
sim_plotrisk(x)
}
\arguments{
\item{x}{output from sim_riskbydistance}
}
\description{
create plot of basic info from sim_riskbydistance output
}
\seealso{
sim_riskbydistance
}
