% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_riskbydistance.R
\name{sim_riskbydistance}
\alias{sim_riskbydistance}
\title{Simulate disparities as risk, pop density, and demographics vary by distance}
\usage{
sim_riskbydistance(
  dist = 1:100,
  ypopdensity.start = 100,
  ypopdensity.end = 50,
  ypopdensity.rate = 0.95,
  ydemog.start = 66,
  ydemog.end = 33,
  ydemog.rate = 0.95,
  yrisk.start = 100,
  yrisk.end = 1,
  yrisk.rate = 0.8,
  silent = FALSE
)
}
\arguments{
\item{dist}{Best not to change this. Distances simulated, such as 1 through 100 (that could be kilometers, for example).}

\item{ypopdensity.start}{population density at closest point to emissions source}

\item{ypopdensity.end}{at furthest point}

\item{ypopdensity.rate}{decay rate with distance such as 0.95 for 5 percent drop per 1 of the 100 distance increments}

\item{ydemog.start}{Demographic group of interest nearby, as percent of pop, 0-100}

\item{ydemog.end}{at furthest point}

\item{ydemog.rate}{decay rate with distance such as 0.95 for 5 percent drop per 1 of the 100 distance increments}

\item{yrisk.start}{Risk metric (e.g., 0 to 100) at nearest point.}

\item{yrisk.end}{at furthest point}

\item{yrisk.rate}{decay rate with distance such as 0.95 for 5 percent drop per 1 of the 100 distance increments}

\item{silent}{whether to print subset of table to console, and suggested plot code}
}
\value{
silently returns data.frame of many indicators, at each of the 100 distances
}
\description{
Creates a simple set of 100 distances and at each distance simulated
  demographics, risk levels, population density and size, etc.
  Useful to explore how those decay with distance and how the
  EJ metrics one might use depend on those factors and domain analyzed.
}
\examples{
  x <- sim_riskbydistance()
  sim_plotrisk(x)
  sim_plotratios(x)
}
