% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.health.url.R
\name{state.health.url}
\alias{state.health.url}
\title{Get URL(s) with State health indicator data from RWJF - ** url scheme obsolete now so needs to be redone}
\usage{
state.health.url(ST = NA, scope = "state", ind = 66, dist = 23,
  char = 87, time = 3, viz = "bar", fstate = NA, locs = NA,
  cmp = "stcmp", open.browser = FALSE)
}
\arguments{
\item{ST}{State abbreviation, FIPS code, or name as character vector}

\item{scope}{Character, default is "state" view in results, but can be "national" view as well}

\item{ind}{Health Indicator code number. Number, default is 66. Possible values: \cr
\itemize{
\item 6=Cancer Incidence: Incidence of breast, cervical, lung and colorectal cancer per 100,000 population; age adjusted
\item 7=Cancer Incidence by Race: Incidence of breast, cervical, lung and colorectal cancer per 100,000 population; age adjusted
\item 10=Chronic Disease Prevalence: asthma/CVD/diabetes,
\item 15=Limited Activity: Average number of days in the previous 30 days when a person indicates their activities are limited due to mental or physical health difficulties,
\item 22=Tobacco taxes: State cigarette excise tax rate ($)
\item 25=Income Inequality (Gini Coefficient)
\item 31=life expectancy= Life expectancy at birth: number of years that a newborn is expected to live if current mortality rates continue to apply
\item 44=Public health funding: Per capita state public health funding
\item 45=Poor/Fair Health: Self-reported health status: percent of adults reporting fair or poor health
\item 65=Premature death: Premature deaths: Average number of years of potential life lost prior to age 75 per 100,000 population
\item 66=Premature Death by Race/Ethnicity: Premature deaths: Average number of years of potential life lost prior to age 75 per 100,000 population
}}

\item{dist}{default is 23. unclear what this controls. 23 or 29 or 0 or 19 possible. dist/char were 0/0 for US totals not by race, and were 19/58 for same by race.}

\item{char}{default is 87. unclear what this controls. 91=?, 119=?, 121=?  58? others possible.}

\item{time}{Code for which years are covered by the data. default is 3, which means 2009-2010. Possible values: \cr
\itemize{
\item 3=2009-2010
\item 22=2010-2011
\item 23=2011-2012
\item 5=2000
\item 10=2005
\item 11=2006
\item 12=2007
\item 13=2008
\item 1=2009
\item 14=2010
\item 4=2011
\item 24=2012
\item OTHERS?
}}

\item{viz}{Type of visualization of data. Default is "bar" and can be line or bar or table.}

\item{fstate}{State number. default is NA}

\item{locs}{Locations to compare. default is NA}

\item{cmp}{Comparisons to view. Default is "stcmp" to see 2+ states compared (or state vs US). Can also see breakdown for one state if set to "brkdwn"}

\item{open.browser}{Logical, default is FALSE. Should URL be opened by launching browser.}
}
\value{
URL(s) character vector, default: \url{http://www.rwjf.org/en/research-publications/research-features/rwjf-datahub/national.html#q/scope/state/ind/66/dist/23/char/87/time/3/viz/bar/fstate/33/locs/33,52,9/cmp/stcmpind/66/dist/23/char/87/time/3/viz/bar/fstate/33/locs/33,52,9/cmp/stcmp}
}
\description{
Robert Woods Johnson Foundation provides health indicator data by state. This function provides a basic interface to those webpages.
}
\details{
NEWER URL SCHEME NOW... E.G. \url{http://www.rwjf.org/en/how-we-work/rel/research-features/rwjf-datahub/national.html#q/scope/national/ind/10/dist/0/char/0/time/3/viz/map/cmp/brkdwn}

see also related percent insured data here: e.g.,
\url{http://datacenter.shadac.org/profile/70#2/alabama/percent,moe,count/a/hide}
DEFAULT IF NO PARAMETERS USED:
state.health.url() \cr
url created: \cr
http://www.rwjf.org/en/research-publications/research-features/rwjf-datahub/national.html#q/scope/state/ind/66/dist/23/char/87/time/3/viz/bar/fstate/33/locs/33,52,9/cmp/stcmp \cr
url that it resolves to on website: \cr
http://www.rwjf.org/en/research-publications/research-features/rwjf-datahub/national.html#q/scope/state/ind/66/dist/23/char/87/time/3/viz/bar/fstate/33/locs/33,52,9/cmp/stcmp \cr
# state.health.url("MD") \cr
# url created: \cr
# \url{http://www.rwjf.org/en/research-publications/research-features/rwjf-datahub/national.html#q/scope/state/ind/66/dist/23/char/87/time/14/viz/bar/fstate/21/locs/21/cmp/stcmp} \cr
# url that it resolves to on website should be  \cr
# \url{http://www.rwjf.org/en/research-publications/research-features/rwjf-datahub/national.html#q/scope/state/ind/66/dist/23/char/87/time/3/viz/bar/fstate/21/locs/21,52/cmp/stcmp} \cr
#	URL scheme for linking to Resources for health data by state or by county: \cr \cr
#State-level data from SHADAC: \url{http://www.shadac.org/} \cr
#State-level data from RWJF: \url{http://www.rwjf.org/en/research-publications/research-features/rwjf-datahub.html} \cr
***** see entire US clickable map of states
# shell.exec('http://www.rwjf.org/en/research-publications/research-features/rwjf-datahub/national.html#q/scope/national/ind/31/dist/29/char/119/time/13/viz/map/fstate/2/locs/2,52/cmp/brkdwn')
# Use this URL format to figure out API or state-specific set of links to nice state reports.
Public health, premature deaths, MD vs US, 2009-2010
\url{http://www.rwjf.org/en/research-publications/research-features/rwjf-datahub/national.html#q/scope/state/ind/66/dist/23/char/91/time/3/viz/bar/fstate/21/locs/21,52/cmp/stcmp}
Public health, life expectancy, MD (state number 21) vs US
\url{http://www.rwjf.org/en/research-publications/research-features/rwjf-datahub/national.html#q/scope/state/ind/31/dist/29/char/119/time/14/viz/line/fstate/21/locs/21,52/cmp/brkdwn}
\url{http://www.rwjf.org/en/research-publications/research-features/rwjf-datahub/national.html#q/scope/state/ind/31/dist/29/char/119/time/14/viz/bar/fstate/21/locs/21,52/cmp/stcmp} \cr
Cancer by race in arizona (defaults to 2008-2009):
\url{http://www.rwjf.org/en/research-publications/research-features/rwjf-datahub/national.html#q/scope/state/ind/7/dist/22/char/85/time/18/viz/bar/fstate/3/locs/3,52/cmp/stcmp}
}
\examples{
#
 shell.exec(state.health.url('CA'))

}
