% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url.censusblock.R
\name{url.censusblock}
\alias{url.censusblock}
\title{See webpage with data on Census block(s)}
\usage{
url.censusblock(
  fips,
  censustable = "P2",
  censusfile = "DEC/10_PL",
  launch = TRUE
)
}
\arguments{
\item{fips}{vector of FIPS}

\item{censustable}{P2 by default but can be another table code. e.g., see \url{http://factfinder.census.gov/faces/affhelp/jsf/pages/metadata.xhtml?lang=en&type=dataset&id=dataset.en.DEC_10_SF1}}

\item{censusfile}{DEC/10_PL by default. Also see DEC_10_SF1 for example.}

\item{launch}{TRUE by default, whether to open page in web browser}
}
\value{
Can open a webpage. Returns vector of URL(s) as character
}
\description{
DRAFT CODE TO see table webpage of BLOCK DATA VIA AFF WITHOUT NEEDING API KEY
}
\details{
For information on FIPS codes, see \url{http://www.census.gov/geo/reference/ansi.html}
and also see \url{https://www.census.gov/geo/reference/geoidentifiers.html}
For links to AFF, see \url{http://factfinder2.census.gov/files/AFF_deep_linking_guide.pdf}
e.g. to get 1 block census 2010 pop, for block fips 360610127001000   \url{http://factfinder2.census.gov/bkmk/table/1.0/en/DEC/10_SF1/P1/1000000US360610127001000}
e.g. to get 1 block census 2010 RACE/ETH/NHWA, for block fips 360610127002001  \url{http://factfinder2.census.gov/bkmk/table/1.0/en/DEC/10_PL/P2/1000000US360610127002001}
Notice the second one gets P2 from 10_PL not 10_SF1, because P2 in SF1 means something different
TO GET RACE/ETHNICITY CENSUS 2010 COUNTS ON ONE BLOCK:
block fips 360610127002001
\url{http://factfinder2.census.gov/bkmk/table/1.0/en/DEC/10_PL/P2/1000000US360610127002001}
TO GET RACE/ETHNICITY CENSUS 2010 COUNTS ON TWO BLOCKS:
\url{http://factfinder2.census.gov/bkmk/table/1.0/en/DEC/10_PL/P2/1000000US360610127002001|1000000US360610127002000}
}
\seealso{
(http://ejanalysis.github.io/countyhealthrankings)
}
