% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.pctiles.by.zone.R
\name{write.pctiles.by.zone}
\alias{write.pctiles.by.zone}
\title{create lookup table as file of percentiles, mean, sd by state or region}
\usage{
write.pctiles.by.zone(mydf, filename, zone.vector)
}
\arguments{
\item{mydf}{data.frame with numeric data. Each column will be examined to calculate mean, sd, and percentiles, for each zone.}

\item{filename}{prefix to use for filename to be saved locally}

\item{zone.vector}{names of states or regions, for example. same length as wts, or rows in mydf}
}
\description{
Given a data.frame, for each column in the data.frame,
this function just returns percentiles, mean, and standard deviation.
Also saves that as a csv file.
}
\seealso{
\link{write.wtd.pctiles.by.zone}
}
