% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.pctiles.R
\name{write.pctiles}
\alias{write.pctiles}
\title{Write csv file lookup table - percentiles, mean, standard deviation}
\usage{
write.pctiles(x, filename)
}
\arguments{
\item{x}{Data.frame, required.}

\item{filename}{Name, or full path and name, of the file to be saved. Required.}
}
\value{
A data.frame with percentiles, mean, and standard deviation. Same number of columns as x had.
}
\description{
Given a data.frame, for each column in the data.frame,
  this function just returns percentiles, mean, and standard deviation.
  Also saves that as a csv file.
}
