% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.RR.tables.R
\name{write.RR.tables}
\alias{write.RR.tables}
\title{Save RR tables to disk, one per zone.}
\usage{
write.RR.tables(my.RR.table, folder = getwd())
}
\arguments{
\item{my.RR.table}{Required RR table from function such as \code{\link[=RR]{RR()}}}

\item{folder}{Optional directory, default is current working directory. Specifies where to save files.}
}
\value{
Returns the file names as a character vector, after saving them locally.
Saves one file per zone, with rownames and header.
Format is one demographic group per row and one environmental indicator per column, plus max per row and max per column
}
\description{
This function breaks up a 3-dimensional table of RR values (e.g., by demographic group, by environmental indicator, by geographic zone),
and saves data for each zone to a 2-dimensional table in a file.
Requires table in format provided by \code{\link[=RR]{RR()}} and related functions.
}
\examples{
  # RRS.REGION <- RR.table(mydat=bg, Enames=names.e,
  # Dnames=c(names.d, names.d.subgroups.pct), popcolname='pop', Zcolname='REGION')
  # write.RR.tables(RRS.REGION)
}
\seealso{
\code{\link[=RR]{RR()}}
}
