% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.wtd.pctiles.by.zone.R
\name{write.wtd.pctiles.by.zone}
\alias{write.wtd.pctiles.by.zone}
\title{create lookup table as file of pop-weighted percentiles, mean, sd by state or region}
\usage{
write.wtd.pctiles.by.zone(mydf, wts, filename, zone.vector)
}
\arguments{
\item{mydf}{data.frame with numeric data. Each column will be examined to calculate mean, sd, and percentiles, for each zone.}

\item{wts}{vector of numbers such as population counts as weights, as long as nrow(mydf)}

\item{filename}{prefix to use for filename to be saved locally}

\item{zone.vector}{names of states or regions, for example. same length as wts, or rows in mydf}
}
\description{
create lookup table as file of pop-weighted percentiles, mean, sd by state or region
}
\examples{
  \dontrun{
  bg = ejscreen::bg21
  pctilevariables <- c(names.e, names.d, names.ej)
   ejanalysis::write.wtd.pctiles(mydf = bg[ , pctilevariables], wts = bg$pop, filename =  'lookupUSA')
   ejanalysis::write.wtd.pctiles.by.zone(mydf = bg[ , pctilevariables], wts = bg$pop,
                                   zone.vector = bg$REGION, filename =  'lookupRegions')
   ejanalysis::write.wtd.pctiles.by.zone(mydf = bg[ , pctilevariables], wts = bg$pop,
                                   zone.vector = bg$ST,     filename =  'lookupStates')
  }

}
