% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addFIPScomponents.R
\name{addFIPScomponents}
\alias{addFIPScomponents}
\title{Add col for each component of longer FIPS
 
 Given a data.frame with FIPS col that is the full state county tract blockgroup FIPS 
 returns the data.frame with extra columns up front, with components of FIPS.}
\usage{
addFIPScomponents(bg)
}
\arguments{
\item{bg}{Data.frame with a character column called FIPS}
}
\value{
Returns the whole data.frame with new columns in front: 
  'FIPS', 'FIPS.TRACT', 'FIPS.COUNTY', 'FIPS.ST', 'ST', 'statename', 'REGION'
}
\description{
Add col for each component of longer FIPS
 
 Given a data.frame with FIPS col that is the full state county tract blockgroup FIPS 
 returns the data.frame with extra columns up front, with components of FIPS.
}
