% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign.pctiles.alt2.R
\name{assign.pctiles.alt2}
\alias{assign.pctiles.alt2}
\title{Assign percentiles to values (alternative formula, and not by zone)}
\usage{
assign.pctiles.alt2(x, weights = NULL, na.rm = TRUE, zone = NULL)
}
\arguments{
\item{x}{vector or data.frame}

\item{weights}{Optional, NULL by default (not fully tested), vector of weights for weighted percentiles (e.g., population weighted).}

\item{na.rm}{Logical, optional, TRUE by default. Should NA values (missing data) be removed first to get percentile of those with valid data.
If FALSE, NA values are treated as being at the high percentiles.}

\item{zone}{Optional, NULL by default, *** not yet implemented here.}
}
\value{
Returns a numeric vector or data.frame same size as x.
}
\description{
For each column look at the distribution of values across all rows,
and find what percentile a given value is at.
}
\details{
Assign percentile as cumulative sum of (the weights ranked by the value x).
Then fixes ties.
# Could also add parameter like in rank(), na.last,
defining na.rm but also where to rank NA values if included, etc.
Default now is like na.last=NA, but like na.last='last' if na.rm=FALSE
Could also add parameter like in rank(), ties.method, defining if ties get min, max, or mean of percentiles initially assigned to ties.
Default for ties right now is like ties.method=max (which might not be what assign.pctiles() does in fact).
}
\examples{
x <- c(30, 40, 50, 12,12,5,5,13,13,13,13,13,8,9,9,9,9,9,10:20,20,20,20,21:30)
weights <- rep(c(2,3), length(x)/2)
cbind(weights, x, PCTILE=assign.pctiles.alt2(x,weights))

# PERCENTILE OF ALL, NOT JUST THOSE WITH VALID DATA, IF na.rm=FALSE,
# but then NA values preclude high percentiles:
x <- c(NA, NA, NA, NA,NA,NA,NA,NA,NA,NA,13,13,8,9,9,9,9,9,10:20,20,20,20,21:30)
weights <- rep(c(2,3), length(x)/2)
cbind(weights, x, PCTILE.alt2=assign.pctiles.alt2(x, weights, na.rm=FALSE),
 pctile=assign.pctiles(x,weights))[order(x),]
cbind(weights, x, PCTILE.alt2=assign.pctiles.alt2(x, weights, na.rm=TRUE),
 pctile=assign.pctiles(x,weights))[order(x),]

V=9
sum(weights[!is.na(x) & x <= V]) / sum(weights[!is.na(x)])

#A value (V) being at this PCTILE\% means that (assuming na.rm=TRUE):

# V >= x  for        PCTILE\% of weights     (for non-NA x), so
# V < x   for 100\% - PCTILE\% of weights     (for non-NA x), or
# PCTILE\% of all weights have V >= x (for non-NA x), so
# 100\% - PCTILE\% of all weights have V < x  (for non-NA x).

x <- c(32, NA, NA, NA,NA,NA,NA,NA,NA,NA,13,13,8,9,9,9,9,9,10:20,20,NA,20,21:30)
weights <- rep(c(2,3), length(x)/2)
cbind(weights, x, PCTILE.alt2=assign.pctiles.alt2(x, weights, na.rm=FALSE),
 pctile=assign.pctiles(x,weights))[order(x),]
cbind(weights, x, PCTILE.alt2=assign.pctiles.alt2(x, weights, na.rm=TRUE),
 pctile=assign.pctiles(x,weights))[order(x),]
}
\seealso{
\code{\link{make.bin.pctile.cols}} to call functions below, converting columns of values to percentiles and then bins \cr
\code{\link{assign.pctiles}} for one vector, assign (weighted) percentile (quantile) to each value within its zone (subset)  \cr
\code{\link{assign.pctiles.alt2}} as an alternative method, to replicate assign.pctiles, but not by zone \cr
\code{\link{get.pctile}} to get (weighted) percentile of just 1+ values within given vector of values \cr
\code{\link{make.pctile.cols}} for a data.frame, assign percentiles, return a same-sized df that is wtd.quantile of each value within its column \cr
\code{\link{make.pctile.cols.alt2}} as an alternative method, to replicate make.pctile.cols \cr
\code{\link{assign.map.bins}} for one vector (or data.frame) of values (e.g. percentiles), return same-sized df that is bin number (map color bin) using preset breaks. \cr
\code{\link{make.bin.cols}} for a data.frame of values (e.g. percentiles), return same-sized df that is bin number (map color bin) using preset breaks. \cr
\code{\link{write.pctiles}} to save file that is lookup table of percentiles for columns of a data.frame \cr
\code{\link{write.pctiles.by.zone}} to save file that is lookup table of percentiles for columns of a data.frame, for each geographic zone (subset of rows) \cr
\code{\link{write.wtd.pctiles}} to save file that is lookup table of weighted percentiles for columns of a data.frame \cr
\code{\link{write.wtd.pctiles.by.zone}} to save file that is lookup table of weighted percentiles for columns of a data.frame, for each geographic zone (subset of rows) \cr
\code{\link{lookup.pctile}} to look up current approx weighted percentiles in a lookup table that is already in global memory \cr
}
