% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgtest.R
\docType{data}
\name{bgtest}
\alias{bgtest}
\title{Test dataset for ejanalysis}
\description{
This data set provides test data, with
Environmental, Demographic, and EJ Index fields,
one row per block group for about 32,698 blockgroups, covering a few states.
This is roughly a subset of ejscreen data like bg21 (see  ejscreen  package).
It does not have of the columns in the ejscreen dataset.
Can be used like this:  data('bgtest', package = 'ejanalysis')
}
\seealso{
\link[ejscreen:names.d]{ejscreen::names.d} (http://ejanalysis.github.io/ejscreen/)
}
