% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get.epa.region.R
\name{get.epa.region}
\alias{get.epa.region}
\title{Identify EPA Region for given state}
\usage{
get.epa.region(state)
}
\arguments{
\item{state}{Vector of numeric or character class, required. Can be state FIPs as number or character, 2-character state abbreviation, or full state name}
}
\value{
Returns a vector of numbers, same length as state
}
\description{
Identify US Environmental Protection Agency region(s) containing given state(s)
}
\details{
For EPA Regions, see \url{http://www2.epa.gov/aboutepa#pane-4} or \url{http://www2.epa.gov/aboutepa/visiting-regional-office}.
  For information on FIPS codes, see \url{http://www.census.gov/geo/reference/ansi.html},
  and also see \url{https://www.census.gov/geo/reference/geoidentifiers.html}
}
\examples{
myregions <- get.epa.region() # to see full list
 myregions <- get.epa.region('DC') # for one state by 2-letter abbreviation
 myregions <- get.epa.region(c('AK', 'NY', 'PR')) # for a vector of 2-letter abbreviations
 myregions <- get.epa.region(c('alaska', 'new york', 'puerto rico')) # for a vector of state names
 myregions <- get.epa.region(c('04', '36', '72')) # for a vector of state FIPS codes as characters with leading zeroes
 myregions <- get.epa.region(c(4,36,72)) # for a vector of state FIPS codes as numeric with no leading zeroes
 myregions <- get.epa.region(c('NY', 'Ohio')) # CANNOT MIX WAYS OF DEFINING STATES - PICK ONE FORMAT
}
\seealso{
\code{\link{clean.fips.1215}}, \code{\link{get.fips.bg}}, \code{\link{get.fips.tract}}, \code{\link{get.fips.county}}, \code{\link{get.fips.st}} to get partial FIPS from longer FIPS, or
  \code{\link{get.name.county}}, \code{\link{get.name.state}} to extract name from longer Census name, or
  \code{\link{get.state.info}}, \code{\link{get.county.info}}, \code{\link{get.epa.region}}
  to look up info such as FIPS, state abbreviation, statename, countyname, or region based on FIPS or name.
}

