% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get.pctile.R
\name{get.pctile}
\alias{get.pctile}
\title{Determine (Weighted) Percentiles}
\usage{
get.pctile(x, values, wts = 1, or.tied = TRUE, na.rm = TRUE)
}
\arguments{
\item{x}{Vector of numeric values (the distribution within which percentiles are sought)}

\item{values}{Required vector of one or more numbers for which percentiles are sought.}

\item{wts}{Optional weights, default is 1 (equal weighting). Useful to find for example what percent of people have an x score (at or) below a given value.}

\item{or.tied}{Default is TRUE which means percentile represents what fraction have x at or below value, not just below value.}

\item{na.rm}{Logical, optional, TRUE by default. Should NA values (missing data) be removed first to get percentile of those with valid (nonmissing) data.}
}
\description{
Given a vector of numbers (a dataset), and one or more specific values of interest,
  determine at what percentile(s) are those selected values. In other words,
  what fraction of all of the numbers (actually weighted based on cumulative sum of weights)
  are smaller than (or tied with) the selected value(s) of interest?
  Based on analyze.stuff::pct.above (\code{\link[analyze.stuff]{pct.above}})
}
\examples{
get.pctile(89:95, 1:100)
get.pctile(89:95, 1:100, c(rep(1,90), rep(10000,10)))
}
\seealso{
\code{\link{make.bin.pctile.cols}} to call functions below, converting columns of values to percentiles and then bins \cr
\code{\link{assign.pctiles}} for one vector, assign (weighted) percentile (quantile) to each value within its zone (subset)  \cr
\code{\link{assign.pctiles.alt2}} as an alternative method, to replicate assign.pctiles, but not by zone \cr
\code{\link{get.pctile}} to get (weighted) percentile of just 1+ values within given vector of values \cr
\code{\link{make.pctile.cols}} for a data.frame, assign percentiles, return a same-sized df that is wtd.quantile of each value within its column \cr
\code{\link{make.pctile.cols.alt2}} as an alternative method, to replicate make.pctile.cols \cr
\code{\link{assign.map.bins}} for one vector (or data.frame) of values (e.g. percentiles), return same-sized df that is bin number (map color bin) using preset breaks. \cr
\code{\link{make.bin.cols}} for a data.frame of values (e.g. percentiles), return same-sized df that is bin number (map color bin) using preset breaks. \cr
\code{\link{write.pctiles}} to save file that is lookup table of percentiles for columns of a data.frame \cr
\code{\link{write.pctiles.by.zone}} to save file that is lookup table of percentiles for columns of a data.frame, for each geographic zone (subset of rows) \cr
\code{\link{write.wtd.pctiles}} to save file that is lookup table of weighted percentiles for columns of a data.frame \cr
\code{\link{write.wtd.pctiles.by.zone}} to save file that is lookup table of weighted percentiles for columns of a data.frame, for each geographic zone (subset of rows) \cr
\code{\link{lookup.pctile}} to look up current approx weighted percentiles in a lookup table that is already in global memory \cr
}

