% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.pctile.R
\name{lookup.pctile}
\alias{lookup.pctile}
\title{Find approx wtd percentiles in lookup table that is in memory}
\usage{
lookup.pctile(myvector, varname.in.lookup.table, lookup, zone)
}
\arguments{
\item{myvector}{Numeric vector, required. Values to look for in the lookup table.}

\item{varname.in.lookup.table}{Character element, required. Name of column in lookup table to look in to find interval where a given element of myvector values is.}

\item{lookup}{Either lookup must be specified, or a lookup table called us must already be in memory. This is the lookup table data.frame with a PCTILE column and column whose name is the value of varname.in.lookup.table.}

\item{zone}{Character element, optional. If specified, must appear in a column called REGION within the lookup table. For example, it could be 'NY' for New York State.}
}
\value{
By default, returns numeric vector length of myvector.
}
\description{
This is used with lookup or us, a data.frame that is a lookup table created by
\code{\link{write.pctiles}}, \code{\link{write.pctiles.by.zone}}, \code{\link{write.wtd.pctiles}}, or \code{\link{write.wtd.pctiles.by.zone}}.
The data.frame us or lookup must have a field called "PCTILE" that contains quantiles/percentiles
and other column(s) with values that fall at those percentiles.
This function accepts lookup table (or uses one called us if that is in memory),
and finds the number in the PCTILE column that corresponds to where a specified value (in myvector)
appears in the column called varname.in.lookup.table.
The function just looks for where the specified value fits between values in the lookup table and returns the approximate percentile as found in the PCTILE column.
}
\examples{
# places.etc$new.pctile.pm <- lookup.pctile(places.etc$countymean, "PM25")
}
\seealso{
\code{\link{write.pctiles}}, \code{\link{write.pctiles.by.zone}}, \code{\link{write.wtd.pctiles}}, and \code{\link{write.wtd.pctiles.by.zone}}
}
