% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.pctile.cols.alt2.R
\name{make.pctile.cols.alt2}
\alias{make.pctile.cols.alt2}
\title{Alternative way to make columns of (weighted) percentiles from columns of values}
\usage{
make.pctile.cols.alt2(
  raw.data.frame,
  weights,
  as.df = TRUE,
  prefix = "pctile.",
  na.rm = TRUE,
  zone = NULL
)
}
\arguments{
\item{raw.data.frame}{Data.frame of values}

\item{weights}{Optional Numeric vector of weights to create weighted percentiles, such as population-weighted quantiles. Unweighted if not specified. Vector same length as number of rows in data.frame.}

\item{as.df}{Optional logical TRUE by default, in which case matrix results are converted to data.frame}

\item{prefix}{Optional character element, default is 'pctile.', provides text to paste to beginning of input data.frame column names to use as output column names.}

\item{na.rm}{Default is TRUE. Passed to \code{\link[=assign.pctiles.alt2]{assign.pctiles.alt2()}}}

\item{zone}{NULL by default, but if a vector is provided, it defines zones to group by, so percentiles are within a given zone only.}
}
\value{
Returns a matrix or data.frame
}
\description{
Takes a data.frame of values and returns a data.frame of percentiles,
showing the percentile of a value within all values in its column.
}
\seealso{
\code{\link{make.bin.pctile.cols}} to call functions below, converting columns of values to percentiles and then bins \cr
\code{\link{assign.pctiles}} for one vector, assign (weighted) percentile (quantile) to each value within its zone (subset)  \cr
\code{\link{assign.pctiles.alt2}} as an alternative method, to replicate assign.pctiles, but not by zone \cr
\code{\link{get.pctile}} to get (weighted) percentile of just 1+ values within given vector of values \cr
\code{\link{make.pctile.cols}} for a data.frame, assign percentiles, return a same-sized df that is wtd.quantile of each value within its column \cr
\code{\link{make.pctile.cols.alt2}} as an alternative method, to replicate make.pctile.cols \cr
\code{\link{assign.map.bins}} for one vector (or data.frame) of values (e.g. percentiles), return same-sized df that is bin number (map color bin) using preset breaks. \cr
\code{\link{make.bin.cols}} for a data.frame of values (e.g. percentiles), return same-sized df that is bin number (map color bin) using preset breaks. \cr
\code{\link{write.pctiles}} to save file that is lookup table of percentiles for columns of a data.frame \cr
\code{\link{write.pctiles.by.zone}} to save file that is lookup table of percentiles for columns of a data.frame, for each geographic zone (subset of rows) \cr
\code{\link{write.wtd.pctiles}} to save file that is lookup table of weighted percentiles for columns of a data.frame \cr
\code{\link{write.wtd.pctiles.by.zone}} to save file that is lookup table of weighted percentiles for columns of a data.frame, for each geographic zone (subset of rows) \cr
\code{\link{lookup.pctile}} to look up current approx weighted percentiles in a lookup table that is already in global memory \cr
}
