% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.dvars.R
\docType{data}
\name{names.dvars}
\alias{names.dvars}
\alias{names.d}
\alias{names.d.subgroups}
\alias{Dlist}
\alias{demographic-variables}
\title{Fieldnames of demographic columns in ejscreen data}
\usage{
data('names.dvars', package = 'ejscreen')
}
\description{
This data set provides variables that hold the colnames
  of demographic fields in data.frames that may be used in the related ejscreen package
  to make it easier to refer to them as a vector, e.g., mydf[ , names.e]
}
\seealso{
\code{\link[ejscreen]{names.dvars}}
}
\keyword{datasets}
