% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/names.ejvars.R
\docType{data}
\name{names.ejvars}
\alias{EJ-variable-names}
\alias{names.ej}
\alias{names.ejvars}
\title{Fieldnames of environmental justice indicator columns in ejanalysis data}
\format{A series of variables (each is a character vector of colnames):
\itemize{
 \item "names.ej"
 \item "names.ej.bin"
 \item "names.ej.burden.eo"
 \item "names.ej.burden.eo.bin"
 \item "names.ej.burden.eo.pctile"
 \item "names.ej.burden.svi6"
 \item "names.ej.burden.svi6.bin"
 \item "names.ej.burden.svi6.pctile"
 \item "names.ej.pct.eo"
 \item "names.ej.pct.eo.bin"
 \item "names.ej.pct.eo.pctile"
 \item "names.ej.pct.svi6"
 \item "names.ej.pct.svi6.bin"
 \item "names.ej.pct.svi6.pctile"
 \item "names.ej.pctile"
 \item "names.ej.svi6"
 \item "names.ej.svi6.bin"
 \item "names.ej.svi6.pctile"
 \item "namesall.ej"
 \item "namesall.ej.bin"
 \item "namesall.ej.pctile"
}}
\source{
Names developed for this package. No external data source.
}
\usage{
data('names.ejvars')
}
\description{
This data set provides variables that hold the colnames
  of environmental indicator fields in data.frames that may be used in the ejanalysis package
  to make it easier to refer to them as a vector, e.g., mydf[ , names.ej]
}
\keyword{datasets}

