% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.evars.R
\docType{data}
\name{names.evars}
\alias{names.evars}
\alias{names.e}
\alias{environmental-variable-names}
\alias{Elist}
\title{Fieldnames of environmental indicator columns in ejscreen data}
\description{
This data set provides variables that hold the colnames
  of environmental indicator fields in data.frames that may be used in the ejscreen package
  to make it easier to refer to them as a vector, e.g., mydf[ , names.e]\cr
  data('names.evars', package = 'ejscreen')
}
\seealso{
\code{\link[ejscreen]{names.evars}}
}
\keyword{datasets}
