% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pop.cdf.density.R
\name{pop.cdf.density}
\alias{pop.cdf.density}
\title{Overlay Two PDFs as Weighted Histogram-Like Density Plots}
\usage{
pop.cdf.density(e, dcount, refcount, etxt, dtxt, brks = 10, ...)
}
\arguments{
\item{e}{Environmental or other indicator values vector}

\item{dcount}{Vector of weights for the demographic group of interest, such as population counts of Hispanics by Census tract.}

\item{refcount}{Vector of weights for the reference group, such as population counts of individuals who are not Hispanic, by Census tract.}

\item{etxt}{Character string to name e in graph}

\item{dtxt}{Character string to name d in graph}

\item{brks}{Default is 10. Passed as breaks param to plot function}

\item{...}{other parameters passed to \code{density}}
}
\value{
Creates a plot
}
\description{
Use plot(density()) to see overlay of two distribution functions.
}
\details{
Assumes you have weights for each and are comparing values in one group vs another.
}
\examples{
\\donotrun{
  pop.cdf.density(e = e, dcount = dcount, refcount = refcount, etxt = etxt, dtxt = dtxt,
}
}
\seealso{
\code{\link{pop.cdf}}   \code{\link{pop.cdf2}} \code{\link{pop.ecdf}}  \code{\link{pop.cdf.density}}
}

