% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RR.plot.R
\name{RR.plot}
\alias{RR.plot}
\title{Draw lineplot comparing RR values by group}
\usage{
RR.plot(RRS, d, e, zone = "NY")
}
\arguments{
\item{RRS}{Table as from \code{\link{RRS}} function, of relative risk by group, risk type, and zone (3 dimensions).}

\item{d}{Demographic group names to be found as names of first dim of RRS}

\item{e}{Environmental factor or risk type names to be found as names of second dim of RRS}

\item{zone}{Zone name to be found among names for third dim of RRS. Default is "NY"}
}
\value{
draws a plot
}
\description{
Draws a plot using relative risk information, one line per group
}
\examples{
#
  ###
}
\seealso{
\code{\link{RR}}
}

