% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RR.table.add.R
\name{RR.table.add}
\alias{RR.table.add}
\title{Merge tables of Relative Risk results}
\usage{
RR.table.add(rrs1, rrs2, zones2)
}
\arguments{
\item{rrs1}{First table from RR.table()}

\item{rrs2}{Another table from RR.table()}

\item{zones2}{Zones in rrs2, as vector, default is dimnames(rrs2)[\link{3}]}
}
\value{
Returns a new array
}
\description{
Merge table of Relative Risk results for some zones with table for USA overall
}
\examples{
 RRS.US  <- RR.table(mydat=bg, Enames=names.e, Dnames=c(names.d, names.d.subgroups.pct),
  popcolname='pop')
 RRS.ST  <- RR.table(mydat=bg, Enames=names.e, Dnames=c(names.d, names.d.subgroups.pct),
  popcolname='pop', Zcolname='ST')
 RRS <- RR.table.add(RRS.ST, RRS.US)
 RRS[ 'pctlowinc', , ]
 RRS[ , , 'CA']
 RRS[ , 'pm', ]
 RR.table.sort(RRS)

 RRS.REGION  <- RR.table(mydat=bg, Enames=names.e, Dnames=c(names.d, names.d.subgroups.pct),
  popcolname='pop', Zcolname='REGION')
 RRS2 <- RR.table.add(RRS, RRS.REGION)
 RRS2[ , , '8']
}
\seealso{
\itemize{
\item \code{\link{ej.indexes}} for local contribution to a variety of overall disparity metrics such as excess risk
\item \code{\link{RR}} to calculate overall disparity metric as relative risk (RR), ratio of mean environmental indicator values across demographic groups
\item \code{\link{RR.table}} to create 3-D table of RR values, by demographic group by environmental indicator by zone
\item \code{\link{RR.table.sort}} to sort existing RR table
\item \code{\link{RR.table.add}} to add zone(s) to existing RR table
\item \code{\link{write.RR.tables}} to write a file with a table or RR by indicator by group
\item \code{\link{ej.added}} to find EJ Index as local contribution to sum of EJ Indexes
\item \code{\link{RR.cut.if.gone}} to find local contribution to RR
\item \code{\link{RR.if.address.top.x}} to find how much RR would change if top-ranked places had different conditions
\item \code{\link{rrfv}} for obsolete attempt to vectorize rrf
\item \code{\link{rrf}} for older simpler function for RR of one indicator for one demographic group
\item \code{\link{pop.ecdf}} to compare plots of cumulative frequency distribution of indicator values by group
}
}
