% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RR.table.add.R
\name{RR.table.add}
\alias{RR.table.add}
\title{Merge tables of Relative Risk results}
\usage{
RR.table.add(rrs1, rrs2, zones2)
}
\arguments{
\item{rrs1}{First table from RR.table()}

\item{rrs2}{Another table from RR.table()}

\item{zones2}{Zones in rrs2, as vector, default is third element of dimnames(rrs2)}
}
\value{
Returns a new array
}
\description{
Merge table of Relative Risk results for some zones with table for USA overall
}
\examples{

 RRS.US  <- RR.table(mydat=bg, Enames=names.e, Dnames=c(names.d, names.d.subgroups.pct),
  popcolname='pop')
 RRS.ST  <- RR.table(mydat=bg, Enames=names.e, Dnames=c(names.d, names.d.subgroups.pct),
  popcolname='pop', Zcolname='ST')
 RRS <- RR.table.add(RRS.ST, RRS.US)
 # RRS[ "pctlowinc", , ]
 # RRS[ , ,  "CA"]
 # RRS[ , "pm", ]
 RR.table.sort(RRS)

 RRS.REGION  <- RR.table(mydat=bg, Enames=names.e, Dnames=c(names.d, names.d.subgroups.pct),
  popcolname='pop', Zcolname='REGION')
 RRS2 <- RR.table.add(RRS, RRS.REGION)
 # RRS2[ , , "8"]

}
\seealso{
\code{\link[=RR]{RR()}}
}
