% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RR.table.R
\name{RR.table}
\alias{RR.table}
\title{Table of Relative Risk results by zone by group by envt risk factor}
\usage{
RR.table(
  mydat,
  Enames = ejscreen::names.e[ejscreen::names.e \%in\% names(mydat)],
  Dnames = ejscreen::names.d[ejscreen::names.d \%in\% names(mydat)],
  popcolname = "pop",
  Zcolname,
  testing = FALSE,
  digits = 4
)
}
\arguments{
\item{mydat}{Data.frame of input data, one row per geographic unit such as US Census block groups, or tracts.}

\item{Enames}{names of columns with environmental risk factor data, default is names.e from ejscreen package}

\item{Dnames}{names of columns with percent (fraction) of population that is in each given demographic group, default is names.d from ejscreen package}

\item{popcolname}{name of column with total population count, default is pop}

\item{Zcolname}{name of column with name of zone such as US State name}

\item{testing}{default is FALSE}

\item{digits}{Default is 4. How many significant digits to use.}
}
\value{
Compiles RR values in array of 3 dimensions: \code{RRS[Dnames, Enames, Zcolnames]}
Returns a matrix with
one demographic group per row,
one environmental risk indicator per column, and
third dimension for which zone (e.g., which US State)
}
\description{
Make table of Relative Risk results by zone by group by envt risk factor.
See source code for notes on this work.
}
\examples{
    # (This is very slow right now)
    # Ratios <- ejanalysis::RR.table(bg21plus, Enames = names.e, Dnames = c(names.d, names.d.subgroups), popcolname = 'pop', digits = 2)
    # MeansByGroup_and_Ratios <- ejanalysis::RR.means(subset(bg21plus, select=names.e), subset(bg21plus, select = c(names.d, names.d.subgroups)), bg21plus$pop)
data(bgtest, package = 'ejanalysis')
RRS.US  <- RR.table(mydat = bgtest, Enames = names.e, Dnames = c(names.d, names.d.subgroups.pct),
 popcolname = 'pop')
RRS.ST  <- RR.table(mydat = bgtest, Enames = names.e, Dnames = c(names.d, names.d.subgroups.pct),
 popcolname = 'pop', Zcolname = 'ST')
RRS <- RR.table.add(RRS.ST, RRS.US)
RRS['pctlowinc', , ]
RRS[ , , 'CA']  # RRS[,,'PR']
RRS[ , 'pm', ]
RRS.REGION  <- RR.table(mydat = bgtest, Enames = names.e, Dnames = c(names.d, names.d.subgroups.pct),
 popcolname='pop', Zcolname='REGION')
RRS2 <- RR.table.add(RRS, RRS.REGION)
RRS2[ , , '8']
}
\seealso{
\itemize{
\item \code{\link{ej.indexes}} for local contribution to a variety of overall disparity metrics such as excess risk
\item \code{\link{RR}} to calculate overall disparity metric as relative risk (RR), ratio of mean environmental indicator values across demographic groups
\item \code{\link{RR.table}} to create 3-D table of RR values, by demographic group by environmental indicator by zone
\item \code{\link{RR.table.sort}} to sort existing RR table
\item \code{\link{RR.table.add}} to add zone(s) to existing RR table
\item \code{\link{write.RR.tables}} to write a file with a table or RR by indicator by group
\item \code{\link{ej.added}} to find EJ Index as local contribution to sum of EJ Indexes
\item \code{\link{RR.cut.if.gone}} to find local contribution to RR
\item \code{\link{RR.if.address.top.x}} to find how much RR would change if top-ranked places had different conditions
\item \code{\link{rrfv}} for obsolete attempt to vectorize rrf
\item \code{\link{rrf}} for older simpler function for RR of one indicator for one demographic group
\item \code{\link{pop.ecdf}} to compare plots of cumulative frequency distribution of indicator values by group
}
}
