% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RR.table.view.R
\name{RR.table.view}
\alias{RR.table.view}
\title{Just print Relative Risks (RR) info by demog group by risk type by zone}
\usage{
RR.table.view(
  x,
  d = dimnames(x)[[1]],
  e = dimnames(x)[[2]],
  zone = dimnames(x)[[3]]
)
}
\arguments{
\item{d}{Optional. Vector of names of demographic groups that must be subset of first dimension of RRS. Default is all.}

\item{e}{Optional. Vector of names of environmental risk indicators that must be subset of second dimension of RRS. Default is all.}

\item{zone}{Optional. Vector of zone names that must be subset of names of third dimension of RRS. Default is all.}

\item{rrs}{A three dimensional array that is created by \code{\link{RR}} and has format like RRS[d, e, zone]}
}
\value{
prints RRS[d, e, zone]
}
\description{
Currently does not do anything other than print. Takes zone-specific RRS values and prints them for viewing.
  Just another way to specify which subset of the RR array you want to view.
}
\examples{
 t(round(RR.table.view(
   RR.table(bgtest, names.e[1:3], names.d[3:4], 'pop', Zcolname = 'statename'),
   d = 'pctlowinc'
  ), 2))
}
\seealso{
\itemize{
\item \code{\link{ej.indexes}} for local contribution to a variety of overall disparity metrics such as excess risk
\item \code{\link{RR}} to calculate overall disparity metric as relative risk (RR), ratio of mean environmental indicator values across demographic groups
\item \code{\link{RR.table}} to create 3-D table of RR values, by demographic group by environmental indicator by zone
\item \code{\link{RR.table.sort}} to sort existing RR table
\item \code{\link{RR.table.add}} to add zone(s) to existing RR table
\item \code{\link{write.RR.tables}} to write a file with a table or RR by indicator by group
\item \code{\link{ej.added}} to find EJ Index as local contribution to sum of EJ Indexes
\item \code{\link{RR.cut.if.gone}} to find local contribution to RR
\item \code{\link{RR.if.address.top.x}} to find how much RR would change if top-ranked places had different conditions
\item \code{\link{rrfv}} for obsolete attempt to vectorize rrf
\item \code{\link{rrf}} for older simpler function for RR of one indicator for one demographic group
\item \code{\link{pop.ecdf}} to compare plots of cumulative frequency distribution of indicator values by group
}
}
