% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RR.table.vs.us.R
\name{RR.table.vs.us}
\alias{RR.table.vs.us}
\title{Relative Risks (RR) in zones as ratios to US values}
\usage{
RR.table.vs.us(rrs, d, e, zone)
}
\arguments{
\item{rrs}{A three dimensional array that is created by \code{\link{RR}} and has format like RRS[d, e, zone] but where zone must include the name USA}

\item{d}{Vector of names of demographic groups that must be subset of first dimension of RRS}

\item{e}{Vector of names of environmental risk indicators that must be subset of second dimension of RRS}

\item{zone}{Vector of zone names that must be subset of names of third dimension of RRS}
}
\value{
numeric results same shape as RRS[d, e, zone]
}
\description{
Takes zone-specific RRS values and divides each value by the corresponding value for the USA overall.
}
\seealso{
\itemize{
\item \code{\link{ej.indexes}} for local contribution to a variety of overall disparity metrics such as excess risk
\item \code{\link{RR}} to calculate overall disparity metric as relative risk (RR), ratio of mean environmental indicator values across demographic groups
\item \code{\link{RR.table}} to create 3-D table of RR values, by demographic group by environmental indicator by zone
\item \code{\link{RR.table.sort}} to sort existing RR table
\item \code{\link{RR.table.add}} to add zone(s) to existing RR table
\item \code{\link{write.RR.tables}} to write a file with a table or RR by indicator by group
\item \code{\link{ej.added}} to find EJ Index as local contribution to sum of EJ Indexes
\item \code{\link{RR.cut.if.gone}} to find local contribution to RR
\item \code{\link{RR.if.address.top.x}} to find how much RR would change if top-ranked places had different conditions
\item \code{\link{rrfv}} for obsolete attempt to vectorize rrf
\item \code{\link{rrf}} for older simpler function for RR of one indicator for one demographic group
\item \code{\link{pop.ecdf}} to compare plots of cumulative frequency distribution of indicator values by group
}
}
