% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrf.R
\name{rrf}
\alias{rrf}
\title{RR for one environmental indicator, one demographic group}
\usage{
rrf(e, d, pop, na.rm = TRUE)
}
\arguments{
\item{e}{Environmental indicator vector, one number per place, required.
e is environmental index or health risk level (e.g., PM2.5 concentration to which this person or place is exposed)}

\item{d}{Demographic indicator vector, required, one number per place, fraction of population that is in group of interest
d is percent of place that is selected demog group (e.g. percent Hispanic) (or d=1 or 0 if this is a vector of individuals)}

\item{pop}{Population total per place, required, of which d is a fraction.
pop is population count of place (or pop=1 if this is a vector of individuals)}

\item{na.rm}{Logical optional TRUE by default in which case NA values (missing values) are removed first. If FALSE, any NA value in pop, e, or d would make result NA.}
}
\value{
Returns one number
}
\description{
**Probably obsolete given \code{\link[=RR]{RR()}}. \cr\cr
Inputs are vectors over small places like Census blocks or block groups
or possibly individuals (ideally) but then d would be a dummy=1 for selected group and 0 for people not in selected group
}
\examples{
#
 # rrf(places$pm, places[ , unlist(Dlist)[1]],  places$pop)
}
\seealso{
\itemize{
\item \code{\link{ej.indexes}} for local contribution to a variety of overall disparity metrics such as excess risk
\item \code{\link{RR}} to calculate overall disparity metric as relative risk (RR), ratio of mean environmental indicator values across demographic groups
\item \code{\link{RR.table}} to create 3-D table of RR values, by demographic group by environmental indicator by zone
\item \code{\link{RR.table.sort}} to sort existing RR table
\item \code{\link{RR.table.add}} to add zone(s) to existing RR table
\item \code{\link{write.RR.tables}} to write a file with a table or RR by indicator by group
\item \code{\link{ej.added}} to find EJ Index as local contribution to sum of EJ Indexes
\item \code{\link{RR.cut.if.gone}} to find local contribution to RR
\item \code{\link{RR.if.address.top.x}} to find how much RR would change if top-ranked places had different conditions
\item \code{\link{rrfv}} for obsolete attempt to vectorize rrf
\item \code{\link{rrf}} for older simpler function for RR of one indicator for one demographic group
\item \code{\link{pop.ecdf}} to compare plots of cumulative frequency distribution of indicator values by group
}
}
