% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterEJ_D_E.R
\name{scatterEJ_D_E}
\alias{scatterEJ_D_E}
\title{scatterplot EJScreen Demog vs Envt vs EJ Index}
\usage{
scatterEJ_D_E(
  d = bg21$pctile.VSI.eo,
  e = bg21$pctile.traffic.score,
  ejbin = bg21$bin.EJ.DISPARITY.traffic.score.eo,
  ename = "",
  legendtitle = "Percentiles of EJ Index (with yellow/orange/red as in the maps)",
  colors11 = c("white", "lightgray", "lightgreen", "darkgray", "purple", "black",
    "darkblue", "lightblue", "yellow", "orange", "red"),
  main = "EJ Index percentile bins",
  mylegend = c("<10", "10-20", "20s", "30s", "40s", "50-60", "60s", "70s", "80-90",
    "90-95", ">95"),
  ...
)
}
\arguments{
\item{d}{vector percentile demographic index}

\item{e}{vector percentile environmental indicator}

\item{ejbin}{vector bin number 1 through N (bins of percentiles) for EJ Index for e}

\item{ename}{friendly name for graphic, optional}

\item{legendtitle}{optional}

\item{colors11}{optional}

\item{main}{optional}

\item{mylegend}{vector of text labels for the N bins}

\item{...}{passed to plot (but not to points for the various colors subsets)}
}
\value{
Nothing. Draws a scatter plot.
}
\description{
scatterplot EJScreen Demog vs Envt vs EJ Index
}
