% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url.open.R
\name{url.open}
\alias{url.open}
\title{Launch a web browser to go to a URL, like browseURL does}
\usage{
url.open(myurl)
}
\arguments{
\item{myurl}{required character element, URL to open}
}
\value{
Just launches browser to open URL
}
\description{
Just the same as \code{\link[=browseURL]{browseURL()}}
}
\seealso{
\code{\link[=browseURL]{browseURL()}}, and \code{\link[=shell.exec]{shell.exec()}} which this uses
}
