% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worstplaces.R
\name{worstplaces}
\alias{worstplaces}
\title{Places with worst RR}
\usage{
worstplaces(rrs, d = "pctlowinc", e = "pm", n = 10, digits = 2)
}
\arguments{
\item{rrs}{Required. This has to be the output of one of the functions like \code{\link[=RR.table]{RR.table()}}}

\item{d}{name of demographic field with percent (fraction) of population that is in each given demographic group, in dimnames(rrs)[\link{1}]}

\item{e}{name of environmental risk factor in dimnames(rrs)[\link{2}]}

\item{n}{worst 10 by default}

\item{digits}{round to 2 by default}
}
\value{
matrix
}
\description{
View one key part of table of Relative Risk results by group by envt risk factor
}
\examples{
 \dontrun{
  data(bgtest, package = 'ejanalysis')
  RRS.US  <- RR.table(mydat = bgtest, Enames = names.e, Dnames = names.d, popcolname = 'pop')
  RRS.ST  <- RR.table(mydat = bgtest, Enames = names.e, Dnames = names.d, popcolname = 'pop',
                  Zcolname = 'ST')
  RRS <- RR.table.add(RRS.ST, RRS.US)
  worstplaces(RRS.US)
}
}
\seealso{
\itemize{
\item \code{\link{ej.indexes}} for local contribution to a variety of overall disparity metrics such as excess risk
\item \code{\link{RR}} to calculate overall disparity metric as relative risk (RR), ratio of mean environmental indicator values across demographic groups
\item \code{\link{RR.table}} to create 3-D table of RR values, by demographic group by environmental indicator by zone
\item \code{\link{RR.table.sort}} to sort existing RR table
\item \code{\link{RR.table.add}} to add zone(s) to existing RR table
\item \code{\link{write.RR.tables}} to write a file with a table or RR by indicator by group
\item \code{\link{ej.added}} to find EJ Index as local contribution to sum of EJ Indexes
\item \code{\link{RR.cut.if.gone}} to find local contribution to RR
\item \code{\link{RR.if.address.top.x}} to find how much RR would change if top-ranked places had different conditions
\item \code{\link{rrfv}} for obsolete attempt to vectorize rrf
\item \code{\link{rrf}} for older simpler function for RR of one indicator for one demographic group
\item \code{\link{pop.ecdf}} to compare plots of cumulative frequency distribution of indicator values by group
}
}
