% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/write.RR.tables.R
\name{write.RR.tables}
\alias{write.RR.tables}
\title{Save RR tables to disk, one per zone.}
\usage{
write.RR.tables(my.RR.table, folder = getwd())
}
\arguments{
\item{my.RR.table}{Required RR table from function such as \code{\link{RR}}}

\item{folder}{Optional directory, default is current working directory. Specifies where to save files.}
}
\value{
Returns the file names as a character vector, after saving them locally.
  Saves one file per zone, with rownames and header.
  Format is one demographic group per row and one environmental indicator per column, plus max per row and max per column
}
\description{
This function breaks up a 3-dimensional table of RR values (e.g., by demographic group, by environmental indicator, by geographic zone),
  and saves data for each zone to a 2-dimensional table in a file.
  Requires table in format provided by \code{\link{RR}} and related functions.
}
\examples{
# RRS.REGION <- RR.table(mydat=bg, Enames=names.e, Dnames=c(names.d, names.d.subgroups.pct), popcolname='pop', Zcolname='REGION')
  # write.RR.tables(RRS.REGION)
}
\seealso{
\itemize{
\item \code{\link{ej.indexes}} for local contribution to a variety of overall disparity metrics such as excess risk
\item \code{\link{RR}} to calculate overall disparity metric as relative risk (RR), ratio of mean environmental indicator values across demographic groups
\item \code{\link{RR.table}} to create 3-D table of RR values, by demographic group by environmental indicator by zone
\item \code{\link{RR.table.sort}} to sort existing RR table
\item \code{\link{RR.table.add}} to add zone(s) to existing RR table
\item \code{\link{write.RR.tables}} to write a file with a table or RR by indicator by group
\item \code{\link{ej.added}} to find EJ Index as local contribution to sum of EJ Indexes
\item \code{\link{RR.cut.if.gone}} to find local contribution to RR
\item \code{\link{RR.if.address.top.x}} to find how much RR would change if top-ranked places had different conditions
\item \code{\link{rrfv}} for obsolete attempt to vectorize rrf
\item \code{\link{rrf}} for older simpler function for RR of one indicator for one demographic group
\item \code{\link{pop.ecdf}} to compare plots of cumulative frequency distribution of indicator values by group
}
}

