% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.wtd.pctiles.by.zone.R
\name{write.wtd.pctiles.by.zone}
\alias{write.wtd.pctiles.by.zone}
\title{create lookup table as file of pop-weighted or unwtd percentiles, mean, sd for US or by state or region}
\usage{
write.wtd.pctiles.by.zone(
  mydf,
  wts = NULL,
  filename = NULL,
  zone.vector = NULL,
  zoneOverallName = "USA"
)
}
\arguments{
\item{mydf}{data.frame with numeric data. Each column will be examined to calculate mean, sd, and percentiles, for each zone.}

\item{wts}{optional vector of numbers such as population counts as weights, as long as nrow(mydf)}

\item{filename}{prefix to use for filename to be saved locally (.csv is added by the function). If not provided, no file is saved.}

\item{zone.vector}{optional names of states or regions, for example. same length as wts, or rows in mydf}

\item{zoneOverallName}{optonal If not by zone, name of entire domain to use in table column called REGION. Default is USA.}
}
\description{
also see \cr\cr
  ejscreen.lookuptables  and \cr\cr
  write.wtd.pctiles.by.zone   and \cr\cr
  table.pop.pctile  and \cr\cr
  map service with lookup tables \cr\cr \cr\cr
\preformatted{
}
}
\examples{
  \dontrun{
  bg = ejscreen::bg21plus # want demog subgroups but also want PR eventually
  pctilevariables <- c(names.e, names.d, names.d.subgroups, names.ej)
   ejanalysis::write.wtd.pctiles(mydf = bg[ , pctilevariables], wts = bg$pop, filename =  'lookupUSA21')
   ejanalysis::write.wtd.pctiles.by.zone(mydf = bg[ , pctilevariables], wts = bg$pop,
                                   zone.vector = bg$REGION, filename =  'lookupRegions21')
   ejanalysis::write.wtd.pctiles.by.zone(mydf = bg[ , pctilevariables], wts = bg$pop,
                                   zone.vector = bg$ST,     filename =  'lookupStates21')
  }

}
